There was an attempt.. it seems the usermod approach won't work
Checking this non working code for the record. The idea is to usermod
to change the uid:gid in entrypoint script. There are two approaches
I took to do this:
1. pass the -u flag to docker run and try to do the usermod. Won't work because we're not in the passwd file aand so user mod fails
2. pass env vars with the host uid:gid you want to run as, and then usermod to change in-flight (fails, as we're trying to change the uid while logged in/running proc

It seems eding the container /etc/passwd file is the only workable approach. One more thing to try is to run eentrypoint.sh before dumbinit - will do that next