WORKDIR no longer required, added support for externally mounting /home/coder
The entrypoint.sh script is now copied to /usr/local/bin, companion
to the code-server binary. The WORKDIR directory is no longer created
as nothing is copied to it. Lastly, we've modified the external volume
mount to export the entire home directory. This will avoid UFS layer
creation for .cache and .local directory writes, which appears to
happen by default.