Support opening multi-folder workspaces
A workspace is described by a JSON file named *.code-workspace in
Visual Studio Code.  Therefore, code-server should let the user
select a file for the "Open Workspace" action.

Also, fix a bug in `workbench` that `URI.file()` can only accept a
string of path, not a raw dump instance of URI object.