replace express-static-gzip with gunzipping ver.
express-static-gzip doesn't gunzip gzipped files for clients that don't
have 'Accept-Encoding: gzip', so it has been replaced by a simple clone
that does the same thing as the original module and also uses zlib to
gunzip the .gz file.

This middleware is compatible with files that aren't .gz in the
rootFolder, although it will not gzip them on the fly (to my knowledge).