fix: use correct command to Open Folder on Welcome page (#3437)
* chore: update CHANGELOG

* fix: use openFolder in welcomePage

In 1.56.2, they modified the WelcomePage use of OpenFileFolder in
src/vs/workbench/contrib/debug/browser/welcomeView.ts to only use on macOS +
Desktop (i.e. not Web). This caused the "Open Folder" to not work on macOS +
macOS-like environments.

Instead, we use OpenFolder which is a registered command and works as expected.