Restore async storage write patch
This issue can be reliably replicated by creating an extension with the
following code:

await context.globalState.update("key", "value")
vscode.commands.executeCommand("workbench.action.reloadWindow");

If you do this the global state will never update. This can be seen by
logging with a notification or browsing the global state manually.

This is probably a bug with upstream because without this patch it is
impossible to wait for the actual storage write.

This actually required more patching than we previously had; I don't
think it was working in 1.56 either.