Restore proposed API patch
Without this it is impossible to enable the proposed API for extensions.

To replicate the issue run a build or search for enableProposedApiForAll
and set it to false (it is set to true during development) then try
running an extension that uses the proposed API like the GitHub pull
request extension and look for an error in the browser console.