Import cli instead of cliProcessMain
cliProcessMain excludes all the code imported by cli since normally cli
imports cliProcessMain.  As a consequence trying to use it directly
results in missing imports.  Call through cli instead.

This does result in parsing arguments twice but it means we have less to
patch and it aligns with how we will ultimately do things.

Remove the process.exit since this cli runs main at the top-level and
does not export anything we can use to wait on it.