PROXY-DOMAIN may be a subdomain + domain
This logging was confusing when a user passes a full <subdomain>.<domain> as proxy-domain, for instance when running code-server with cloudflare tunnels.
This is likely not the best fix. If preferred, I can do some work here to determine if the string is just a domain, or a subdomain+domain, and switch logging behavior based on that condition.