refactor: duplicate onLine to test/utils/helpers
Unfortunately the way Playwright `requires()` files means you have to be
careful with what you import in your tests.

There is a function import related to getting the `package.json` from
the `lib/vscode` directory in the same file where `onLine` is defined.
But since we don't checkout submodules for e2e tests, this file doesn't
exist.

This move duplicates the code to a file so we don't have that
problem/warning from Playwright.