fix: add fake package.json in test-e2e
One of our helper functions `getPackageJson` logs a warning if it can't
find the `package.json` for the path provided. Because we don't clone
submodules in CI running the e2e tests, this warning logs and makes it
hard to read e2e test output.

By creating a fake one if needed in our test-e2e.sh script, we prevent
this warning from being logged.